split out data components, fix private/shared status for admin users
