+++ eventkit_cloud/ui/static/ui/app/components/CustomTableRow.js
@@ -27,12 +27,10 @@ export class CustomTableRow extends Component {
                 display: 'flex',
                 flex: '1 1 auto',
                 flexWrap: 'wrap',
+                alignItems: 'center',
                 backgroundColor: colors.secondary,
                 color: colors.text_primary,
-                paddingRight: '10px',
-                paddingBottom: '0px',
-                paddingTop: '10px',
-                paddingLeft: '10px',
+                padding: '10px',
                 wordBreak: 'break-word',
                 width: '100%',
                 ...this.props.dataStyle,
